wacom: move the Protocol enum to here
This isn't a config setting, it's something we really need to handle in the
wacom communication, the config setting is just a side-effect
